chore: harden v1 lifecycle polish
