+++ crates/bookforge-cli/src/commands/translate/reporting.rs
@@ -1,0 +1,69 @@
+use anyhow::Result;
+use bookforge_core::{
+    config::TranslationConfig,
+    segment::{BlockTranslation, Segment},
+};
+use bookforge_epub::rebuild_epub;
+use bookforge_llm::{QaSegmentReview, SegmentTranslation};
+use bookforge_store::{JobRecord, JobStore};
+
+use crate::{
+    cost::estimate_cost_usd,
+    report::{ReportInput, write_report},
+};
+
+pub fn block_translations(translations: &[SegmentTranslation]) -> Vec<BlockTranslation> {
+    translations
+        .iter()
+        .flat_map(|translation| translation.blocks.iter().cloned())
+        .collect()
+}
+
+pub fn print_summary_rebuild_and_report(
+    store: &JobStore,
+    job: &JobRecord,
+    book: &bookforge_core::ir::Book,
+    segments: &[Segment],
+    translations: &[SegmentTranslation],
+    qa_reviews: &[QaSegmentReview],
+    config: &TranslationConfig,
+) -> Result<()> {
+    let block_translations = block_translations(translations);
+    rebuild_epub(book, &block_translations, &config.output)?;
+    let summary = store
+        .summary(&job.id)?
+        .ok_or_else(|| anyhow::anyhow!("job '{}' was not found after translation", job.id))?;
+    let segment_records = store.segment_records(&job.id)?;
+    let report = write_report(ReportInput {
+        job,
+        summary: &summary,
+        segments,
+        segment_records: &segment_records,
+        translations,
+        qa_reviews,
+        output: &config.output,
+    })?;
+
+    println!(
+        "Translated: {}/{} segments",
+        summary.succeeded, summary.total_segments
+    );
+    println!("Cached: {}", summary.cached);
+    println!("Retried: {}", summary.retried);
+    println!("Needs review: {}", summary.needs_review);
+    println!("Failed: {}", summary.failed);
+    println!("Input tokens: {}", summary.input_tokens);
+    println!("Output tokens: {}", summary.output_tokens);
+    if let Some(cost) = estimate_cost_usd(
+        &job.provider,
+        &job.model,
+        summary.input_tokens,
+        summary.output_tokens,
+    ) {
+        println!("Estimated cost: ${cost:.6}");
+    }
+    println!("Output: {}", config.output.display());
+    println!("Report: {}", report.markdown.display());
+
+    Ok(())
+}
